unauthorized in process

+++ server/src/main/java/service/LogoutService.java
@@ -1,6 +1,7 @@
 package service;
 
-import dataAccess.*;
+import dataAccess.exception.DataAccessException;
+import dataAccess.object.protocol.AuthDAO;
 import result.MessageResult;
 
 public class LogoutService {
@@ -11,6 +12,9 @@ public class LogoutService {
     }
     public MessageResult logout(String authToken) {
         try {
+            if (authToken == null || authToken.isEmpty()) {
+                return new MessageResult("Error: unauthorized");
+            }
             authDAO.deleteAuth(authToken);
             return new MessageResult(null);
         } catch (DataAccessException e) {
